autooptimize: extract shared FakeZeroClawSupervisor and consolidate _enable_pma imports (-113 LOC)
