+++ tests/pma_support/managed_threads.py
@@ -128,3 +128,118 @@ def install_fake_supervisor(
     app.state.app_server_supervisor = supervisor
     app.state.app_server_events = object()
     return supervisor
+
+
+class FakeZeroClawSupervisor:
+    def __init__(
+        self,
+        *,
+        sequential_sessions: bool = False,
+        block_on_wait: bool = False,
+    ) -> None:
+        self.create_calls: list[tuple[Path, str | None]] = []
+        self.attach_calls: list[tuple[Path, str]] = []
+        self.turn_calls: list[dict[str, object]] = []
+        self.wait_started: list[str] = []
+        self._session_seq = 0
+        self._sequential_sessions = sequential_sessions
+        self._block_on_wait = block_on_wait
+        self._session_events: dict[str, asyncio.Event] = {}
+
+    async def create_session(
+        self, workspace_root: Path, title: str | None = None
+    ) -> str:
+        self.create_calls.append((workspace_root, title))
+        if self._sequential_sessions:
+            self._session_seq += 1
+            session_id = f"zeroclaw-session-{self._session_seq}"
+        else:
+            session_id = "zeroclaw-session-1"
+        self._session_events[session_id] = asyncio.Event()
+        return session_id
+
+    async def attach_session(self, workspace_root: Path, session_id: str) -> str:
+        self.attach_calls.append((workspace_root, session_id))
+        return session_id
+
+    async def start_turn(
+        self,
+        workspace_root: Path,
+        conversation_id: str,
+        prompt: str,
+        *,
+        model: str | None = None,
+    ) -> str:
+        self.turn_calls.append(
+            {
+                "workspace_root": workspace_root,
+                "conversation_id": conversation_id,
+                "prompt": prompt,
+                "model": model,
+            }
+        )
+        if self._sequential_sessions:
+            return f"{conversation_id}-turn-{len(self.turn_calls)}"
+        return f"zeroclaw-turn-{len(self.turn_calls)}"
+
+    async def wait_for_turn(
+        self,
+        workspace_root: Path,
+        conversation_id: str,
+        turn_id: str,
+        *,
+        timeout: float | None = None,
+    ):
+        _ = workspace_root, turn_id, timeout
+        if self._block_on_wait:
+            self.wait_started.append(conversation_id)
+            await self._session_events[conversation_id].wait()
+            return type(
+                "Result",
+                (),
+                {
+                    "status": "ok",
+                    "assistant_text": f"zeroclaw-output:{conversation_id}",
+                    "raw_events": [],
+                    "errors": [],
+                },
+            )()
+        return type(
+            "Result",
+            (),
+            {
+                "status": "ok",
+                "assistant_text": f"zeroclaw-output-{len(self.turn_calls)}",
+                "raw_events": [],
+                "errors": [],
+            },
+        )()
+
+    async def stream_turn_events(
+        self, workspace_root: Path, conversation_id: str, turn_id: str
+    ):
+        _ = workspace_root, conversation_id, turn_id
+        if False:
+            yield None
+
+
+class FakeZeroClawEventSupervisor:
+    def __init__(
+        self,
+        *,
+        events: list[dict[str, str]] | None = None,
+        error: Exception | None = None,
+    ) -> None:
+        self._events = events or []
+        self._error = error
+
+    async def list_turn_events(
+        self, workspace_root: Path, session_id: str, turn_id: str
+    ) -> list[dict[str, str]]:
+        _ = workspace_root, session_id, turn_id
+        if self._error is not None:
+            raise self._error
+        return list(self._events)
+
+    async def list_turn_events_by_turn_id(self, turn_id: str) -> list[dict[str, Any]]:
+        return await self.list_turn_events(Path("."), "zeroclaw-session-1", turn_id)
