+++ tests/surfaces/web/routes/pma_routes/test_tail_stream_characterization.py
@@ -1,24 +1,13 @@
 from __future__ import annotations
 
-import json
-from pathlib import Path
-
 import pytest
 from fastapi.testclient import TestClient
-from tests.conftest import write_test_config
+from tests.pma_support import _enable_pma
 
-from codex_autorunner.core.config import CONFIG_FILENAME, DEFAULT_HUB_CONFIG
 from codex_autorunner.core.pma_thread_store import PmaThreadStore
 from codex_autorunner.server import create_hub_app
 
 pytestmark = pytest.mark.slow
-
-
-def _enable_pma(hub_root: Path) -> None:
-    cfg = json.loads(json.dumps(DEFAULT_HUB_CONFIG))
-    cfg.setdefault("pma", {})
-    cfg["pma"]["enabled"] = True
-    write_test_config(hub_root / CONFIG_FILENAME, cfg)
 
 
 class TestManagedThreadStatusShape:
