+++ tests/test_pma_managed_threads_interrupt.py
@@ -1,25 +1,16 @@
 from __future__ import annotations
 
-import json
 from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
 
-from codex_autorunner.core.config import CONFIG_FILENAME, DEFAULT_HUB_CONFIG
 from codex_autorunner.core.orchestration.runtime_bindings import (
     clear_runtime_thread_binding,
 )
 from codex_autorunner.core.pma_thread_store import PmaThreadStore
 from codex_autorunner.server import create_hub_app
-from tests.conftest import write_test_config
-
-
-def _enable_pma(hub_root: Path) -> None:
-    cfg = json.loads(json.dumps(DEFAULT_HUB_CONFIG))
-    cfg.setdefault("pma", {})
-    cfg["pma"]["enabled"] = True
-    write_test_config(hub_root / CONFIG_FILENAME, cfg)
+from tests.pma_support import _enable_pma
 
 
 @pytest.mark.slow
