+++ tests/test_pma_managed_threads_turns.py
@@ -1,23 +1,14 @@
 from __future__ import annotations
 
-import json
 from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
 
-from codex_autorunner.core.config import CONFIG_FILENAME, DEFAULT_HUB_CONFIG
 from codex_autorunner.server import create_hub_app
-from tests.conftest import write_test_config
+from tests.pma_support import _enable_pma
 
 pytestmark = pytest.mark.slow
-
-
-def _enable_pma(hub_root: Path) -> None:
-    cfg = json.loads(json.dumps(DEFAULT_HUB_CONFIG))
-    cfg.setdefault("pma", {})
-    cfg["pma"]["enabled"] = True
-    write_test_config(hub_root / CONFIG_FILENAME, cfg)
 
 
 def _repo_owner(hub_env) -> dict[str, str]:
